Refactored common.html to post both action name and value. Also made a workaround to ensure this works with the slider. I think we're ready for a pull request...
