+++ core/src/main/java/org/mindtrails/domain/tracking/ActionLog/Action.java
@@ -1,0 +1,11 @@
+package org.mindtrails.domain.tracking.ActionLog;
+
+import lombok.Getter;
+
+public class Action {
+    @Getter
+    private String name;
+
+    @Getter
+    protected String value;
+}
