+++ core/src/main/java/org/mindtrails/persistence/ActionLogRepository.java
@@ -1,8 +1,7 @@
 package org.mindtrails.persistence;
 
-import org.mindtrails.domain.tracking.ActionLog;
+import org.mindtrails.domain.tracking.ActionLog.ActionLog;
 import org.mindtrails.domain.Participant;
-import org.mindtrails.domain.tracking.ActionLog;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
